Fix warnings in rustc 1.46.0
Values that can return errors must be handled.  Using the `?`
operator will bubble up those errors.

The case of making a closure to pass to `onload` could not use
the `?` in its body, so this uses `expect()` intead.

Also makes constant GENERIC_BOX_STYLES all caps, as the
compiler warned about this as well.  While doing so, made the
string multi-line for legibility.